refactor: chatroom 및 chatting 도메인 구조로 변경
