--- src/main/java/space/space_spring/dao/chat/custom/ChatRoomRepositoryCustom.java
@@ -1,11 +1,0 @@
-package space.space_spring.dao.chat.custom;
-
-import space.space_spring.entity.ChatRoom;
-import space.space_spring.entity.Space;
-import space.space_spring.entity.User;
-
-import java.util.List;
-
-public interface ChatRoomRepositoryCustom {
-    List<ChatRoom> findByUserAndSpace(User user, Space space);
-}
