+++ src/main/java/space/space_spring/domain/chat/chatroom/model/response/ReadChatRoomResponse.java
@@ -1,0 +1,20 @@
+package space.space_spring.domain.chat.chatroom.model.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+@Builder
+public class ReadChatRoomResponse {
+    private List<ChatRoomResponse> chatRoomList;
+
+    public static ReadChatRoomResponse of(List<ChatRoomResponse> chatRooms) {
+        return ReadChatRoomResponse.builder()
+                .chatRoomList(chatRooms)
+                .build();
+    }
+}
