+++ src/test/java/space/space_spring/repository/ChatRoomRepositoryTest.java
@@ -5,12 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import space.space_spring.config.QueryDslConfig;
-import space.space_spring.dao.chat.ChatRoomRepository;
-import space.space_spring.dao.chat.UserChatRoomRepository;
-import space.space_spring.entity.ChatRoom;
+import space.space_spring.domain.chat.chatroom.repository.ChatRoomRepository;
+import space.space_spring.domain.chat.chatroom.repository.UserChatRoomRepository;
+import space.space_spring.domain.chat.chatroom.model.ChatRoom;
 import space.space_spring.entity.Space;
 import space.space_spring.entity.User;
-import space.space_spring.entity.UserChatRoom;
+import space.space_spring.domain.chat.chatting.model.UserChatRoom;
 import space.space_spring.entity.enumStatus.BaseStatusType;
 import space.space_spring.entity.enumStatus.UserSignupType;
