+++ src/test/java/space/space_spring/service/ChatRoomServiceTest.java
@@ -11,12 +11,12 @@
 //import space.space_spring.dao.chat.ChatRoomDao;
 //import space.space_spring.dao.chat.ChattingDao;
 //import space.space_spring.dao.chat.UserChatRoomDao;
-//import space.space_spring.dto.chat.request.CreateChatRoomRequest;
-//import space.space_spring.dto.chat.request.JoinChatRoomRequest;
-//import space.space_spring.dto.chat.response.ChatSuccessResponse;
-//import space.space_spring.dto.chat.response.CreateChatRoomResponse;
-//import space.space_spring.dto.chat.response.ReadChatRoomMemberResponse;
-//import space.space_spring.dto.chat.response.ReadChatRoomResponse;
+//import space.space_spring.domain.chat.chatroom.model.request.CreateChatRoomRequest;
+//import space.space_spring.domain.chat.chatroom.model.request.JoinChatRoomRequest;
+//import space.space_spring.domain.chat.chatroom.model.response.ChatSuccessResponse;
+//import space.space_spring.domain.chat.chatroom.model.response.CreateChatRoomResponse;
+//import space.space_spring.domain.chat.chatroom.model.response.ReadChatRoomMemberResponse;
+//import space.space_spring.domain.chat.chatroom.model.response.ReadChatRoomResponse;
 //import space.space_spring.entity.*;
 //import space.space_spring.entity.enumStatus.UserSignupType;
 //import space.space_spring.entity.enumStatus.UserSpaceAuth;
